move basemodel class to validation_pro
